requirements.txt
boto3
requests
pyyaml
python-jose
